Fix error behavior during download (#196)

+++ ssds/blobstore/local.py
@@ -14,7 +14,7 @@ def catch_blob_not_found(func):
         try:
             return func(self, *args, **kwargs)
         except FileNotFoundError as ex:
-            raise BlobNotFoundError(f"Could not find {self.key}") from ex
+            raise BlobNotFoundError(f"Could not find {self.url}") from ex
     return wrapper
 
 class LocalBlobStore(BlobStore):
@@ -73,8 +73,9 @@ class LocalBlob(Blob):
         if self.url != src_blob.url:
             shutil.copyfile(src_blob._path, self._path)
 
-    @catch_blob_not_found
     def download(self, path: str):
+        if not os.path.isfile(self._path):
+            raise BlobNotFoundError(f"Could not find {self.url}")
         if self.url != path:
             shutil.copyfile(self._path, path)
